Add files via upload
Agent-based simulation in cooperation with the municipal housing company of Potsdam, ProPotsdam GmbH, as field application of the PaSyMo tools. With a research focus on residential mobility and population dynamics in the Schlaatz neighborhood of Potsdam the effects of the refurbishment of a late-GDR social housing district from 2020 to 2035 can be explored. The co-designed interactive model includes various data sets for visualization and interactive exploration in scenario development and is conceived to be explored on an interactive simulation table.